+++ Patrullero/src/frsf/cidisi/exercise/patrullero/search/modelo/ObstaculoParcial.java
@@ -1,9 +1,9 @@
 package frsf.cidisi.exercise.patrullero.search.modelo;
 
 public class ObstaculoParcial extends Obstaculo {
-	private Long retardoMultiplicativo;
+	private Integer retardoMultiplicativo;
 
-	public ObstaculoParcial(Long id, NombreObstaculo nombre, Long tiempoInicio, Long tiempoFin, Visibilidad visibilidad, Lugar lugar, Long retardoMultiplicativo) {
+	public ObstaculoParcial(Long id, NombreObstaculo nombre, Long tiempoInicio, Long tiempoFin, Visibilidad visibilidad, Lugar lugar, Integer retardoMultiplicativo) {
 		super(id, nombre, tiempoInicio, tiempoFin, visibilidad, lugar);
 		this.retardoMultiplicativo = retardoMultiplicativo;
 	}
@@ -14,7 +14,7 @@ public class ObstaculoParcial extends Obstaculo {
 	}
 
 	@Override
-	public Long getPeso(Long peso) {
+	public Integer getPeso(Integer peso) {
 		return peso * retardoMultiplicativo;
 	}
 }
